refactor: new coinjoin/common module with utils code
This code can be used in both server side code and in wallet code
