+++ src/coinjoin/coinjoin.h
@@ -5,6 +5,8 @@
 #ifndef BITCOIN_COINJOIN_COINJOIN_H
 #define BITCOIN_COINJOIN_COINJOIN_H
 
+#include <coinjoin/common.h>
+
 #include <core_io.h>
 #include <netaddress.h>
 #include <primitives/block.h>
@@ -12,7 +14,6 @@
 #include <sync.h>
 #include <timedata.h>
 #include <univalue.h>
-#include <util/ranges.h>
 #include <util/translation.h>
 #include <version.h>
 
@@ -20,7 +21,6 @@
 #include <optional>
 #include <utility>
 
-class CCoinJoin;
 class CConnman;
 class CBLSPublicKey;
 class CBlockIndex;
@@ -110,26 +110,6 @@ public:
     {
         READWRITE(obj.nSessionID, obj.nState, obj.nStatusUpdate, obj.nMessageID);
     }
-};
-
-/** Holds a mixing input
- */
-class CTxDSIn : public CTxIn
-{
-public:
-    // memory only
-    CScript prevPubKey;
-    bool fHasSig{false}; // flag to indicate if signed
-    int nRounds{-10};
-
-    CTxDSIn(const CTxIn& txin, CScript script, int nRounds) :
-        CTxIn(txin),
-        prevPubKey(std::move(script)),
-        nRounds(nRounds)
-    {
-    }
-
-    CTxDSIn() = default;
 };
 
 class CCoinJoinAccept
@@ -357,139 +337,31 @@ public:
     bool GetQueueItemAndTry(CCoinJoinQueue& dsqRet) LOCKS_EXCLUDED(cs_vecqueue);
 };
 
-// helper class
-class CCoinJoin
+// Various helpers and dstx manager implementation
+namespace CCoinJoin
 {
-private:
-    // make constructor, destructor and copying not available
-    CCoinJoin() = default;
-    ~CCoinJoin() = default;
-    CCoinJoin(CCoinJoin const&) = delete;
-    CCoinJoin& operator=(CCoinJoin const&) = delete;
+    extern Mutex cs_mapdstx;
 
-    // static members
-    static constexpr std::array<CAmount, 5> vecStandardDenominations{
-            (10 * COIN) + 10000,
-            (1 * COIN) + 1000,
-            (COIN / 10) + 100,
-            (COIN / 100) + 10,
-            (COIN / 1000) + 1,
-    };
-
-    static Mutex cs_mapdstx;
-    static std::map<uint256, CCoinJoinBroadcastTx> mapDSTX GUARDED_BY(cs_mapdstx);
-
-    static void CheckDSTXes(const CBlockIndex* pindex, const llmq::CChainLocksHandler& clhandler) LOCKS_EXCLUDED(cs_mapdstx);
-
-public:
-    static constexpr std::array<CAmount, 5> GetStandardDenominations() { return vecStandardDenominations; }
-    static constexpr CAmount GetSmallestDenomination() { return vecStandardDenominations.back(); }
-
-    static constexpr bool IsDenominatedAmount(CAmount nInputAmount) { return AmountToDenomination(nInputAmount) > 0; }
-    static constexpr bool IsValidDenomination(int nDenom) { return DenominationToAmount(nDenom) > 0; }
-
-    /*
-        Return a bitshifted integer representing a denomination in vecStandardDenominations
-        or 0 if none was found
-    */
-    static constexpr int AmountToDenomination(CAmount nInputAmount)
-    {
-        for (size_t i = 0; i < vecStandardDenominations.size(); ++i) {
-            if (nInputAmount == vecStandardDenominations[i]) {
-                return 1 << i;
-            }
-        }
-        return 0;
-    }
-
-    /*
-        Returns:
-        - one of standard denominations from vecStandardDenominations based on the provided bitshifted integer
-        - 0 for non-initialized sessions (nDenom = 0)
-        - a value below 0 if an error occurred while converting from one to another
-    */
-    static constexpr CAmount DenominationToAmount(int nDenom)
-    {
-        if (nDenom == 0) {
-            // not initialized
-            return 0;
-        }
-
-        size_t nMaxDenoms = vecStandardDenominations.size();
-
-        if (nDenom >= (1 << nMaxDenoms) || nDenom < 0) {
-            // out of bounds
-            return -1;
-        }
-
-        if ((nDenom & (nDenom - 1)) != 0) {
-            // non-denom
-            return -2;
-        }
-
-        CAmount nDenomAmount{-3};
-
-        for (size_t i = 0; i < nMaxDenoms; ++i) {
-            if (nDenom & (1 << i)) {
-                nDenomAmount = vecStandardDenominations[i];
-                break;
-            }
-        }
-
-        return nDenomAmount;
-    }
-
-    /*
-    Same as DenominationToAmount but returns a string representation
-    */
-    static std::string DenominationToString(int nDenom);
-
-    static bilingual_str GetMessageByID(PoolMessage nMessageID);
+    bilingual_str GetMessageByID(PoolMessage nMessageID);
 
     /// Get the minimum/maximum number of participants for the pool
-    static int GetMinPoolParticipants();
-    static int GetMaxPoolParticipants();
+    int GetMinPoolParticipants();
+    int GetMaxPoolParticipants();
 
-    static constexpr CAmount GetMaxPoolAmount() { return COINJOIN_ENTRY_MAX_SIZE * vecStandardDenominations.front(); }
+    constexpr CAmount GetMaxPoolAmount() { return COINJOIN_ENTRY_MAX_SIZE * vecStandardDenominations.front(); }
 
     /// If the collateral is valid given by a client
-    static bool IsCollateralValid(CTxMemPool& mempool, const CTransaction& txCollateral);
-    static constexpr CAmount GetCollateralAmount() { return GetSmallestDenomination() / 10; }
-    static constexpr CAmount GetMaxCollateralAmount() { return GetCollateralAmount() * 4; }
+    bool IsCollateralValid(CTxMemPool& mempool, const CTransaction& txCollateral);
 
-    static constexpr bool IsCollateralAmount(CAmount nInputAmount)
-    {
-        // collateral input can be anything between 1x and "max" (including both)
-        return (nInputAmount >= GetCollateralAmount() && nInputAmount <= GetMaxCollateralAmount());
-    }
+    void AddDSTX(const CCoinJoinBroadcastTx& dstx) LOCKS_EXCLUDED(cs_mapdstx);
+    CCoinJoinBroadcastTx GetDSTX(const uint256& hash) LOCKS_EXCLUDED(cs_mapdstx);
 
-    static constexpr int CalculateAmountPriority(CAmount nInputAmount)
-    {
-        if (auto optDenom = ranges::find_if_opt(GetStandardDenominations(), [&nInputAmount](const auto& denom) {
-            return nInputAmount == denom;
-        })) {
-            return (float)COIN / *optDenom * 10000;
-        }
-        if (nInputAmount < COIN) {
-            return 20000;
-        }
+    void UpdatedBlockTip(const CBlockIndex* pindex, const llmq::CChainLocksHandler& clhandler, const CMasternodeSync& mn_sync);
+    void NotifyChainLock(const CBlockIndex* pindex, const llmq::CChainLocksHandler& clhandler, const CMasternodeSync& mn_sync);
 
-        //nondenom return largest first
-        return -1 * (nInputAmount / COIN);
-    }
-
-
-    static void AddDSTX(const CCoinJoinBroadcastTx& dstx) LOCKS_EXCLUDED(cs_mapdstx);
-    static CCoinJoinBroadcastTx GetDSTX(const uint256& hash) LOCKS_EXCLUDED(cs_mapdstx);
-
-    static void UpdatedBlockTip(const CBlockIndex* pindex, const llmq::CChainLocksHandler& clhandler, const CMasternodeSync& mn_sync);
-    static void NotifyChainLock(const CBlockIndex* pindex, const llmq::CChainLocksHandler& clhandler, const CMasternodeSync& mn_sync);
-
-    static void TransactionAddedToMempool(const CTransactionRef& tx) LOCKS_EXCLUDED(cs_mapdstx);
-    static void BlockConnected(const std::shared_ptr<const CBlock>& pblock, const CBlockIndex* pindex) LOCKS_EXCLUDED(cs_mapdstx);
-    static void BlockDisconnected(const std::shared_ptr<const CBlock>& pblock, const CBlockIndex*) LOCKS_EXCLUDED(cs_mapdstx);
-private:
-    static void UpdateDSTXConfirmedHeight(const CTransactionRef& tx, std::optional<int> nHeight);
+    void TransactionAddedToMempool(const CTransactionRef& tx) LOCKS_EXCLUDED(cs_mapdstx);
+    void BlockConnected(const std::shared_ptr<const CBlock>& pblock, const CBlockIndex* pindex) LOCKS_EXCLUDED(cs_mapdstx);
+    void BlockDisconnected(const std::shared_ptr<const CBlock>& pblock, const CBlockIndex*) LOCKS_EXCLUDED(cs_mapdstx);
 };
 
 #endif // BITCOIN_COINJOIN_COINJOIN_H
